스프링게시판 기능추가
추가 구현 기능
6. 로그아웃
7. 글 작성, 답변 시 본인 아이디 입력되어 있도록
8. 본인 아이디로 작성된 게시글만 수정, 삭제할 수 있게

+++ All4_Back/Java_Spirng/Spring18/src/main/java/com/all/spring/command/board/BDeleteCommand.java
@@ -3,6 +3,7 @@ package com.all.spring.command.board;
 import java.util.Map;
 
 import javax.servlet.http.HttpServletRequest;
+import javax.servlet.http.HttpSession;
 
 import org.springframework.ui.Model;
 
@@ -16,9 +17,26 @@ public class BDeleteCommand implements Command {
 		Map<String, Object> map = model.asMap();
 		HttpServletRequest request = (HttpServletRequest) map.get("request");
 		int bId = Integer.parseInt(request.getParameter("bId"));
+		String bName = request.getParameter("bName");
 		
-		BDao dao = new BDao();
-		dao.delete(bId);
+		// HttpSession 객체 가져오기
+	    HttpSession session = request.getSession();  
+	    // HttpSession 객체에서 "user_id" 가져오기
+	    String user_id = (String) session.getAttribute("user_id");  
+		System.out.println("2" + user_id);
+		
+		if (user_id.equals(bName)) {
+			BDao dao = new BDao();
+			dao.delete(bId);
+			
+			model.addAttribute("msg", "삭제함!");
+	        model.addAttribute("url", "/list");
+		}else {
+			model.addAttribute("msg", "남의 글 삭제 안됨!");
+	        model.addAttribute("url", "/list");
+		}
+		
+		
 	}
 
 }
